feat: 프로필 배경 문제 수정 및 좋아요 하다가 말았음

+++ src/main/resources/static/js/reply/reply.js
@@ -1,37 +1,38 @@
 
 
 // 게시글 삭제 기능
-document.querySelector('.delete-button').addEventListener('click', ()=>{
-    let articleId = document.getElementById('articleId').value;
-    let check = confirm("정말로 삭제하시겠습니까?");
-    if (check == false)
-        return;
-    fetch("/article?id=" + articleId, {
-        method: "delete"})
-        .then((response)=>{
-            if(response.status == 200) { //성공
-                let str = response.text().then((res)=>{
-                    alert("글을 삭제하였습니다");
-                    location.href = res;
-                })
-            }
-            else if(response.status == 500) { //로그인 안했을 경우
-                alert("로그인이 필요합니다");
-            }
-            else if(response.status == 400) { // 유효성 미통과
-                response.text().then((res)=>{
-                    let list = JSON.parse(res);
-                    console.log(list);
-                    let error = '';
-                    list['arrayList'].forEach((e)=>{
-                        error += e['message'] + '\n';
+if (document.querySelector('.delete-button') != null) {
+    document.querySelector('.delete-button').addEventListener('click', ()=>{
+        let articleId = document.getElementById('articleId').value;
+        let check = confirm("정말로 삭제하시겠습니까?");
+        if (check == false)
+            return;
+        fetch("/article?id=" + articleId, {
+            method: "delete"})
+            .then((response)=>{
+                if(response.status == 200) { //성공
+                    let str = response.text().then((res)=>{
+                        alert("글을 삭제하였습니다");
+                        location.href = res;
+                    })
+                }
+                else if(response.status == 500) { //로그인 안했을 경우
+                    alert("로그인이 필요합니다");
+                }
+                else if(response.status == 400) { // 유효성 미통과
+                    response.text().then((res)=>{
+                        let list = JSON.parse(res);
+                        console.log(list);
+                        let error = '';
+                        list['arrayList'].forEach((e)=>{
+                            error += e['message'] + '\n';
+                        });
+                        alert(error);
                     });
-                    alert(error);
-                });
-            }
-        })
-})
-
+                }
+            })
+    })
+}
 
 // 댓글 생성 기능
 function submitReply(button) {
